+++ Program.cs
@@ -11,7 +11,7 @@ namespace Test
     {
         static void Main(string[] args)
         {
-// считываем значения и инициализируем массив
+// Считываем значения и инициализируем массив
             string[] str=File.ReadAllLines("input.txt");
             int N = Int32.Parse(str[0]);
             for(int i=0;i<N;i++)
@@ -32,7 +32,7 @@ namespace Test
                         client[i].setType("creditor");
                         client[i].creditor.SetFamilyName(words[1]);
                         client[i].creditor.SetDate(words[2]);
-                        client[i].creditor.SetLoanAmount(ushort.Parse(words[3]));
+                        client[i].creditor.SetLoanAmount(int.Parse(words[3]));
                         client[i].creditor.SetLoanPercentage(float.Parse(words[4]));
                         client[i].creditor.SetDebt(ushort.Parse(words[5]));
                         break;
@@ -40,7 +40,7 @@ namespace Test
                         client[i].setType("contributer");
                         client[i].contributer.SetFamilyName(words[1]);
                         client[i].contributer.SetDate(words[2]);
-                        client[i].contributer.SetDepositAmount(ushort.Parse(words[3]));
+                        client[i].contributer.SetDepositAmount(int.Parse(words[3]));
                         client[i].contributer.SetDepositPercentage(float.Parse(words[4]));
                         break;
                     case "Организация":
@@ -55,6 +55,8 @@ namespace Test
 
                 }
             }
+
+// Выводим массив в консоль
             for(int i=0;i<N;i++)
             {
                 client[i].Print();
